# server/src/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Car API",
        "description": "API was created for the project to car manage and their service requests. The endpoints scheduled below were implemented to create the application as part of the \"Tworzenie serwisów web 2.0\" course.",
        "version": "1.0.0"
    },
    "paths": {
        "/cars": {
            "get": {
                "summary": "Retrieve all cars objects stored in database.",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/Car"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "message": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "Cars"
                ]
            },
            "post": {
                "summary": "Create a new car instance.",
                "parameters": [
                    {
                        "name": "car",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Car"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "message": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "message": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "Cars"
                ]
            }
        },
        "/cars/{id}": {
            "get": {
                "summary": "Retrieve a single car object by its ID.",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the car object.",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/Car"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "string"
                        }
                    }
                },
                "tags": [
                    "Cars"
                ]
            },
            "put": {
                "summary": "Update a car object by ID.",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of car object to update.",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "car",
                        "in": "body",
                        "description": "New car object to replace the old one.",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Car"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "304": {
                        "description": "Not Modified",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                },
                "tags": [
                    "Cars"
                ]
            },
            "delete": {
                "summary": "Remove a car object by ID",
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "description": "ID of the car object to remove",
                        "schema": {
                            "type": "string",
                            "format": "ObjectId"
                        }
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Accepted",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "string"
                        }
                    }
                },
                "tags": [
                    "Cars"
                ]
            }
        },
...
    },
    "definitions": {
        "Car": {
            "type": "object",
            "properties": {
                "_id": {
                    "type": "string",
                    "format": "ObjectId"
                },
                "vin": {
                    "type": "string"
                },
                "brand": {
                    "type": "string"
                },
                "model": {
                    "type": "string"
                },
                "year_of_production": {
                    "type": "integer"
                },
                "task_type": {
                    "type": "string",
                    "enum": [
                        "Repairing",
                        "Varnishing",
                        "Cleaning"
                    ]
                },
                "description": {
                    "type": "string"
                },
                "status": {
                    "type": "string",
                    "enum": [
                        "Pending",
                        "Servicing",
                        "Finished"
                    ]
                }
            },
            "required": [
                "vin",
                "brand",
                "model",
                "year_of_production",
                "task_type",
                "description",
                "status"
            ]
        },
...
    }
}